requirements.txt
pygetwindow
pyautogui
pillow
mss
